+++ gcc/testsuite/gfortran.dg/g77_intrinsics_funcs.f
@@ -1,7 +1,7 @@
 ! { dg-do compile }
 ! Testing g77 intrinsics as subroutines
-      integer*8 i8
-      integer*4 i4
+      integer(kind=8) i8
+      integer i4
       integer i
       character*80 c
